Rename files to use imported ideas
